Updated complicated example so that it works better out of the box
Added a couple lines of code to get the actual path to the caches directory within the app sandbox and then append "/CleanroomLogger" so that this section of code is more plug 'n play. With these couple lines of code, it makes it easier for a new user to CleanroomLogger to copy and paste and get going. Obviously, the previous path does not work out of the box because it is not the full path to the app's sandbox. Plus, it may encourage users to put logs in the Caches directory in case an iPhone user is low on space and the logs are taking up a lot of room.
Also, fixed the formatting of the RotatingLogFIleConfiguration initializer because the last few parameters were off the screen.